feat(inputs.dpdk): merging identifySockets and maintainConnections. Minor fixes
